Separate typings for node and browser
