+++ package.json
@@ -17,39 +17,40 @@
   ],
   "main": "node/index.js",
   "module": "node/module.js",
-  "typings": "render.d.ts",
+  "typings": "src/node.d.ts",
   "esnext": "src/node.js",
   "browser": {
     "node/index.js": "./browser/index.js",
     "node/module.js": "./browser/module.js",
+    "src/node.d.ts": "./src/browser.d.ts",
     "src/node.js": "./src/browser.js"
   },
   "devDependencies": {
-    "@babel/core": "7.0.0-beta.44",
-    "@babel/plugin-transform-react-jsx": "7.0.0-beta.44",
-    "@babel/preset-env": "7.0.0-beta.44",
-    "@babel/register": "7.0.0-beta.44",
-    "@types/node": "^6.0",
+    "@babel/core": "7.0.0-beta.52",
+    "@babel/plugin-transform-react-jsx": "7.0.0-beta.52",
+    "@babel/preset-env": "7.0.0-beta.52",
+    "@babel/register": "7.0.0-beta.52",
+    "@types/node": "^6.0.114",
     "babel-core": "^7.0.0-0",
-    "babel-jest": "^22.4.3",
+    "babel-jest": "^23.2.0",
     "benchr": "^4.3.0",
     "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-airbnb": "^17.0.0",
     "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.10.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-prettier": "^2.5.0",
-    "eslint-plugin-react": "^7.7.0",
-    "fs-extra": "^5.0.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.0",
+    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-react": "^7.10.0",
+    "fs-extra": "^6.0.1",
     "husky": "^0.14.3",
     "hyperapp": "^1.2.6",
-    "jest": "^22.4.3",
-    "prettier": "^1.11.1",
-    "rollup": "^0.57.1",
-    "rollup-plugin-babel": "4.0.0-beta.2",
-    "rollup-plugin-commonjs": "^9.1.0",
+    "jest": "^23.3.0",
+    "prettier": "^1.13.7",
+    "rollup": "^0.62.0",
+    "rollup-plugin-babel": "4.0.0-beta.7",
+    "rollup-plugin-commonjs": "^9.1.3",
     "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-uglify": "^3.0.0",
+    "rollup-plugin-uglify": "^4.0.0",
     "typescript": "^2.9.2"
   },
   "peerDependencies": {
@@ -58,7 +59,7 @@
   },
   "scripts": {
     "lint": "eslint benchmark src test tools",
-    "test": "jest --coverage --no-cache && tsc -p test/ts",
+    "test": "jest --coverage && tsc -p test/ts",
     "build": "node tools/build",
     "precommit": "node tools/precommit",
     "benchmark": "benchr benchmark/benchmark.js --require @babel/register"
