+++ test/ts/browser.tsx
@@ -1,5 +1,5 @@
-import { h, ActionsType, View } from "hyperapp"
-import { renderer, renderToString } from "../../src/index"
+import { h, app, ActionsType, View } from "hyperapp"
+import { withRender, App, Render, renderToString } from "../../src/browser"
 
 namespace Counter {
   export interface State {
@@ -31,17 +31,18 @@ const view: View<Counter.State, Counter.Actions> = (state, actions) => (
   </main>
 )
 
-const counterRendererString = renderer(
-  view,
-  Counter.state,
-  Counter.actions
-)(Infinity)
-
 const counterString = renderToString(
   view,
   Counter.state,
   Counter.actions
 )
 
-console.log(counterRendererString)
+const counterRender = withRender<App<Counter.State, Counter.Actions>, Render<Counter.State, Counter.Actions>>(app)(
+  Counter.state,
+  Counter.actions,
+  view,
+  document.body
+)
+
 console.log(counterString)
+console.log(counterRender.toString())
